facial_recognition_app/requirements.txt
onnxruntime-gpu
opencv-contrib-python
faiss-gpu
h5py
tqdm
numpy
pillow
PyQt5
